Deployment/uxo_classifier_api/requirements.txt
flask==3.1.0
torch==2.6.0
torchvision==0.21.0
pillow==11.2.1
flask-cors==5.0.1
gunicorn==23.0.0 